⚙️ Restructure project
