+++ core/domain/src/main/java/one/reevdev/carserve/core/domain/feature/service/model/ServiceParam.kt
@@ -1,8 +1,8 @@
-package one.reevdev.carserve.core.domain.model.service
+package one.reevdev.carserve.core.domain.feature.service.model
 
 import android.graphics.Bitmap
 import one.reevdev.carserve.core.common.data.emptyString
-import one.reevdev.carserve.core.domain.model.vehicle.VehicleParam
+import one.reevdev.carserve.core.domain.feature.vehicle.model.VehicleParam
 
 data class ServiceParam(
     val symptoms: String = emptyString(),
