+++ core/domain/src/main/java/one/reevdev/carserve/core/domain/feature/service/usecase/ServiceUseCase.kt
@@ -1,9 +1,9 @@
-package one.reevdev.carserve.core.domain.usecase.service
+package one.reevdev.carserve.core.domain.feature.service.usecase
 
 import kotlinx.coroutines.flow.Flow
 import one.reevdev.carserve.core.common.data.Result
-import one.reevdev.carserve.core.domain.model.service.ServiceAnalysis
-import one.reevdev.carserve.core.domain.model.service.ServiceParam
+import one.reevdev.carserve.core.domain.feature.service.model.ServiceAnalysis
+import one.reevdev.carserve.core.domain.feature.service.model.ServiceParam
 
 interface ServiceUseCase {
     fun analyzeService(param: ServiceParam): Flow<Result<ServiceAnalysis>>
