+++ core/domain/src/main/java/one/reevdev/carserve/core/domain/feature/vehicle/model/Transmission.kt
@@ -1,4 +1,4 @@
-package one.reevdev.carserve.core.domain.model.vehicle
+package one.reevdev.carserve.core.domain.feature.vehicle.model
 
 enum class Transmission(val value: String) {
     AUTOMATIC("Automatic"), MANUAL("Manual"), CVT("CVT")
