+++ feature/vehicle/src/main/java/one/reevdev/carserve/vehicle/component/SelectableList.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import one.reevdev.carserve.core.domain.model.vehicle.Transmission
+import one.reevdev.carserve.core.domain.feature.vehicle.model.Transmission
 import one.reevdev.carserve.feature.common.ui.component.LabelText
 import one.reevdev.carserve.feature.common.ui.theme.CarServeTheme
 import one.reevdev.carserve.vehicle.R
