+++ feature/vehicle/src/main/java/one/reevdev/carserve/vehicle/navigation/VehicleNavigation.kt
@@ -3,7 +3,7 @@ package one.reevdev.carserve.vehicle.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import one.reevdev.carserve.core.domain.model.vehicle.VehicleParam
+import one.reevdev.carserve.core.domain.feature.vehicle.model.VehicleParam
 import one.reevdev.carserve.vehicle.screen.AddVehicleRouter
 
 fun NavController.navigateToAddToCar() {
