+++ feature/vehicle/src/main/java/one/reevdev/carserve/vehicle/screen/AddVehicleScreen.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import one.reevdev.carserve.core.common.data.emptyString
-import one.reevdev.carserve.core.domain.model.vehicle.Transmission
-import one.reevdev.carserve.core.domain.model.vehicle.VehicleParam
+import one.reevdev.carserve.core.domain.feature.vehicle.model.Transmission
+import one.reevdev.carserve.core.domain.feature.vehicle.model.VehicleParam
 import one.reevdev.carserve.feature.common.ui.theme.CarServeTheme
 import one.reevdev.carserve.vehicle.R
 import one.reevdev.carserve.vehicle.component.SelectableList
